refactor: 댓글 서비스 테스트를 단위/통합 테스트로 분리 (#81)
기존 CommentServiceImplTest가 단위 테스트와 통합 테스트의 역할을 겸하면서 발생하던 문제를 해결

문제점
•@SpringBootTest로 작성된 CommentServiceImplTest가 다른 통합 테스트(ClubRestClientTest 등)와 컨텍스트 설정이 달라
'All Tests' 실행 시 컨텍스트 재시작 및 로그 오염으로 인한 테스트 실패를 유발했습니다.

•순수 로직을 검증해야 할 단위 테스트에 Spring AOP 기능(@Retryable) 테스트와 로그 출력 검증 로직이 포함되어
테스트가 무겁고 깨지기 쉬운(brittle) 문제가 있었습니다.

변경 내용
1.CommentServiceImplTest를 순수 단위 테스트로 전환
•@SpringBootTest를 제거하고 @ExtendWith(MockitoExtension.class)를 사용하여 Spring 컨텍스트 의존성을 완전히 제거했습니다.
•로그 출력 검증(CapturedOutput)을 제거하고, verify를 사용한 행위 검증에 집중하도록 수정했습니다.

2.CommentServiceIntegrationTest 신규 생성
•@Retryable과 같이 Spring 프레임워크 기능에 의존하는 테스트를 위해 새로운 서비스 계층 통합 테스트 파일을 생성했습니다.
•기존 CommentServiceImplTest에 있던 @Retryable 테스트 케이스를 이곳으로 이전하여 테스트 커버리지 공백을 방지했습니다.
